requirements.deploy.txt


#  Core data 
numpy==2.2.5
pandas==2.3.3
scipy==1.15.3

#  ML 
scikit-learn==1.7.0
xgboost==2.1.1
shap==0.49.1
joblib==1.5.1
lightgbm==4.6.0

# Deep learning (CPU)
torch==2.10.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

#  Time series / econometrics 
statsmodels==0.14.6
arch==8.0.0
pmdarima==2.1.1
MAPIE==1.3.0

#  NLP 
transformers==5.3.0
sentencepiece==0.2.1
tokenizers==0.22.2

#  Financial data 
yfinance==1.2.0
fredapi==0.5.2

#  API framework 
fastapi==0.135.1
uvicorn==0.41.0
starlette==0.52.1
pydantic==2.12.5
pydantic_core==2.41.5
slowapi==0.1.9
python-multipart==0.0.9

# HF Hub
huggingface_hub==1.6.0
safetensors==0.7.0
hf-xet==1.3.2
fsspec==2025.12.0
filelock==3.20.0

#  LLM 
anthropic==0.100.0
httpx==0.28.1
httpcore==1.0.9
anyio==4.12.1
sniffio==1.3.1
jiter==0.14.0

#  Utilities 
python-dotenv==1.2.2
requests==2.32.5
tqdm==4.67.3
packaging==24.2
typing_extensions==4.15.0
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
urllib3==2.6.3
regex==2026.2.28
sympy==1.14.0
networkx==3.4.2
